requirements.txt
Flask==2.3.3
python-telegram-bot==13.15
gspread==5.10.0
oauth2client==4.1.3
